major updates

+++ routes/blogs.js
@@ -1,18 +1,53 @@
 const express = require('express');
 const fetchUser = require('../middleware/fetchUser');
 const Blog = require("../models/Blog");
+const User = require("../models/User");
 const {body, validationResult} = require("express-validator");
 const router = express.Router();
 
 
-// ROUTE 1: get all the Blogs of the user using GET: "/api/blogs/getallblogs" . no Log in required...
+// Get all the Blogs with query using GET: "/api/blogs/getallblogs" .
+// queries {
+// limit: no of docs required
+// tags: tag of the query
+// user: name of the creator
+// }
+// no Log in required...
 router.get('/getallblogs', async (req, res) => {
 
-    const blogs = await Blog.find();
+    const query = {}
+    
+    // Limit
+    const limit = req.query.limit ? req.query.limit : 0
+    
+    // Tags
+    const tags = req.query.tags
+    if (tags) {
+        query.tags = { $all: tags.split(",") }
+    }
+    
+    // Username
+    const userName = req.query.user
+    if (userName) {
+        const users = await User.find({ displayName: userName })
+        const userIds = users.map(user=>user.id)
+
+        if (userIds.length!==0) {
+            query.user = { $in: userIds }
+        }
+    }
+
+    const blogs = await Blog.find(query).limit(limit);
     res.json(blogs);
 })
 
-// ROUTE 2: get blog by id using GET: "/api/blogs/getblog" . no log in required...
+// Get all the Blogs of the particular user using GET: "/api/blogs/getuserblogs" . no Log in required...
+router.get('/getuserblogs', fetchUser, async (req, res)=> {
+    const blogs = await Blog.find({ user: req.user.id })
+    return res.send(blogs)
+})
+
+// Get blog by id using GET: "/api/blogs/getblog" . no log in required...
 router.get('/getblog/:slg', async (req, res) => {
 
     const blog = await Blog.findOne({ slug: req.params.slg })
@@ -24,21 +59,21 @@ router.get('/getblog/:slg', async (req, res) => {
     }
 })
 
-// ROUTE 3: get recent blogs by id using GET: "/api/blogs/recents" . no log in required...
+// Get recent blogs by id using GET: "/api/blogs/recents" . no log in required...
 router.get('/recents', async (req, res) => {
 
     const blogs = await Blog.find({}).sort({date_modified : -1}).limit(3);
     res.json(blogs);
 })
 
-// ROUTE 4: get popular blogs on the basis of likes by id using GET: "/api/blogs/popular" . no log in required...
+// Post popular blogs on the basis of likes by id using GET: "/api/blogs/popular" . no log in required...
 router.get('/populars', async (req, res) => {
 
     const blogs = await Blog.find({}).sort({likes : -1}).limit(6);
     res.json(blogs);
 })
 
-// ROUTE 5: add a blog using POST: "/api/blogs/addblog". Log in required...
+// Post a blog using POST: "/api/blogs/addblog". Log in required...
 router.post('/addblog', fetchUser, [
 
     // Validation array
@@ -58,6 +93,7 @@ router.post('/addblog', fetchUser, [
 
         // Create blog...
         const blog = await Blog.create({
+            user: req.user.id,
             title : req.body.title,
             content : req.body.content,
             description : req.body.description,
@@ -74,7 +110,7 @@ router.post('/addblog', fetchUser, [
     }
 })
 
-// ROUTE 6: edit a blog using PUT: "/api/blogs/editblog/:id". Log in required...
+// Edit a blog using PUT: "/api/blogs/editblog/:id". Log in required...
 router.put('/editblog/:id', fetchUser, [
 
     // Validation Array
@@ -112,12 +148,18 @@ router.put('/editblog/:id', fetchUser, [
     }
 })
 
-// ROUTE 7: delete a blog using DELETE: "/api/blogs/deleteblog/:id". Log in required...
+// Delete a blog using DELETE: "/api/blogs/deleteblog/:id". Log in required...
 router.delete('/deleteblog/:id', fetchUser, async (req, res) => {
-    
-    await Blog.findByIdAndDelete(req.params.id);
-    res.send("blog has been deleted")
-    
+
+    const blog = await Blog.findById(req.params.id);
+
+    if (blog.user==req.user.id) {
+
+        await Blog.findByIdAndDelete(req.params.id);
+        res.send("blog has been deleted")
+        
+    } else res.send("Access Denied...!")
+
 })
